modularize mock functions and test config setup
